feat(fault_proof): implement `handle_game_defense` for proposer (#402)
* feat(fault_proof): implement `handle_game_defense` for proposer

* test: add `test_proposer_defends_successfully` test

* cleanup

* docs: remove link to rs files

* docs: update proposer.md

* apply comments

* apply comment
